Change `Cache-Control` rule
This is needed for proxies that interfere with the request (such as caching). This was the case with the `create-react-app`, when proxying requests to the dev-server backend to your real backend. See this issue: https://github.com/facebook/create-react-app/issues/1633#issuecomment-291353470 . Maybe this header isn't always desired, but I didn't know a good way to make this configurable other than passing a 3rd argument to this `pipe` function, or maybe an extra parameter to the constructor allowing users to overwrite some of these headers if they want.